+++ src/app/tools/tutorial-link/tutorial-link.component.spec.ts
@@ -1,0 +1,48 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { TutorialLinkComponent } from './tutorial-link.component';
+import {HttpClientTestingModule} from '@angular/common/http/testing';
+import {RouterTestingModule} from '@angular/router/testing';
+import {AngularFireModule} from '@angular/fire';
+import {COMMON_CONFIG} from '../../../../test/test-config';
+import {AngularFireAuth} from '@angular/fire/auth';
+import {AngularFirestore} from '@angular/fire/firestore';
+import {FIRESTORESTUB} from '../../../../test/firestore-stub';
+import {PharosApiService} from '../../pharos-services/pharos-api.service';
+import {Apollo} from 'apollo-angular';
+import {ActivatedRoute} from '@angular/router';
+import {MOCKACTIVATEDROUTE} from '../../../../test/mock-activate-route';
+
+describe('TutorialLinkComponent', () => {
+  let component: TutorialLinkComponent;
+  let fixture: ComponentFixture<TutorialLinkComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [
+        HttpClientTestingModule,
+        RouterTestingModule,
+        AngularFireModule.initializeApp(COMMON_CONFIG)
+      ],
+      providers: [
+        AngularFireAuth,
+        { provide: AngularFirestore, useValue: FIRESTORESTUB },
+        PharosApiService,
+        Apollo,
+        { provide: ActivatedRoute, useValue: MOCKACTIVATEDROUTE }
+      ],
+      declarations: [ TutorialLinkComponent ]
+    })
+    .compileComponents();
+  });
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(TutorialLinkComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});